+++ gatsby-node.js
@@ -2,6 +2,7 @@
 // Licensed under the MIT License
 
 exports.createPages = require('./util/gatsby/createPages');
-exports.onCreateNode = require('./util/gatsby/onCreateNode');
+// exports.onCreateNode = require('./util/gatsby/onCreateNode');
 exports.onCreatePage = require('./util/gatsby/onCreatePage');
-exports.onPostBuild = require('./util/gatsby/onPostBuild');
+// TODO Uncomment when the functionality is implemented again.
+// exports.onPostBuild = require('./util/gatsby/onPostBuild');