+++ src/components/link/LocalizedLink.jsx
@@ -7,10 +7,10 @@ import { Link, useStaticQuery, graphql } from 'gatsby';
 import LocalizedAuthorLink from './LocalizedAuthorLink';
 import LocalizedBlogLink from './LocalizedBlogLink';
 import LocalizedCategoryLink from './LocalizedCategoryLink';
+import LocalizedPageLink from './LocalizedPageLink';
 
 import { AUTHOR_SLUG } from '../../constants';
 
-import markdownPageSlugs from '../../../data/markdown-page-slugs.json';
 import pageSlugs from '../../data/page-slugs.json';
 
 const pageKeySlashIndex = 1;
@@ -23,7 +23,7 @@ const createLocalizedSlug = (site, locale, slug) => {
   if (pageKey) {
     return pageKey.split('/').reduce(
       (previous, current) => {
-        const slugs = current in pageSlugs ? pageSlugs : markdownPageSlugs;
+        const slugs = pageSlugs;
         return `${previous}/${slugs[current][locale]}`;
       },
       locale === defaultLocale ? '' : `/${locale}`,
@@ -97,6 +97,9 @@ const LocalizedLink = (props) => {
       }
       case 'ContentfulCategory': {
         return <LocalizedCategoryLink {...props} />;
+      }
+      case 'ContentfulPage': {
+        return <LocalizedPageLink {...props} />;
       }
       default:
         break;