+++ src/templates/index.jsx
@@ -3,7 +3,6 @@
 
 import React from 'react';
 import { graphql } from 'gatsby';
-import { getImage } from 'gatsby-plugin-image';
 import styled from 'styled-components';
 import { useIntl } from 'react-intl';
 import {
@@ -75,48 +74,40 @@ const Centered = styled.div`
 const Page = (props) => {
   const i = createIntl(useIntl());
 
-  const {
-    coverMarkdownRemark: cover,
-    storyMarkdownRemark: story,
-    serviceWebsitesMarkdownRemark: websites,
-    serviceDesignMarkdownRemark: design,
-    serviceEventsMarkdownRemark: events,
-    sectionContactMarkdownRemark: contact,
-    authorAnttiKiviMarkdownRemark: anttiKivi,
-    authorGurmannSainiMarkdownRemark: gurmannSaini,
-    authorAnssiMoilanenMarkdownRemark: anssiMoilanen,
-    imageAnttiKivi,
-    imageGurmannSaini,
-    imageAnssiMoilanen,
-  } = props.data;
+  const { contentfulIndexPage: page } = props.data;
 
   return (
-    <LayoutIndex
-      title={i('indexTitle')}
-      lang={props.pageContext.lang}
-      pageKey={props.pageContext.key}
-    >
-      <IndexCover title={cover.frontmatter.title} htmlTitle>
-        <div dangerouslySetInnerHTML={{ __html: cover.html }} />
+    <LayoutIndex title={page.title} lang={props.pageContext.lang} pageKey={props.pageContext.key}>
+      <IndexCover title={page.introTitle} htmlTitle>
+        <div dangerouslySetInnerHTML={{ __html: page.introBody.childMarkdownRemark.html }} />
       </IndexCover>
-      <StoryCover title={story.frontmatter.title}>
-        <div dangerouslySetInnerHTML={{ __html: story.html }} />
+      <StoryCover title={page.storyTitle}>
+        <div dangerouslySetInnerHTML={{ __html: page.storyBody.childMarkdownRemark.html }} />
       </StoryCover>
       <Break color={'peach'} mode={1} />
       <Section lesserMargin>
-        <H2 id={i('indexServicesId')}>{i('indexServicesTitle')}</H2>
+        <H2 id={page.productsId}>{page.productsTitle}</H2>
         <Cards>
-          <FeatureCard
-            title={websites.frontmatter.title}
-            icon={<DeviceDesktopIcon size={'large'} />}
-          >
-            <div dangerouslySetInnerHTML={{ __html: websites.html }} />
+          <FeatureCard title={page.products[0].title} icon={<DeviceDesktopIcon size={'large'} />}>
+            <div
+              dangerouslySetInnerHTML={{
+                __html: page.products[0].description.childMarkdownRemark.html,
+              }}
+            />
           </FeatureCard>
-          <FeatureCard title={design.frontmatter.title} icon={<PencilIcon size={'large'} />}>
-            <div dangerouslySetInnerHTML={{ __html: design.html }} />
+          <FeatureCard title={page.products[1].title} icon={<PencilIcon size={'large'} />}>
+            <div
+              dangerouslySetInnerHTML={{
+                __html: page.products[1].description.childMarkdownRemark.html,
+              }}
+            />
           </FeatureCard>
-          <FeatureCard title={events.frontmatter.title} icon={<CalendarIcon size={'large'} />}>
-            <div dangerouslySetInnerHTML={{ __html: events.html }} />
+          <FeatureCard title={page.products[2].title} icon={<CalendarIcon size={'large'} />}>
+            <div
+              dangerouslySetInnerHTML={{
+                __html: page.products[2].description.childMarkdownRemark.html,
+              }}
+            />
           </FeatureCard>
         </Cards>
       </Section>
@@ -125,34 +116,19 @@ const Page = (props) => {
         <Icon>
           <NorthStarIcon size={'large'} />
         </Icon>
-        <H2 id={i('indexPortfolioId')}>{i('indexPortfolioTitle')}</H2>
+        <H2 id={page.portfolioId}>{page.portfolioTitle}</H2>
       </Section>
       <Break color={'peach'} mode={3} />
       <Section lesserMargin>
         <Icon>
           <PaperAirplaneIcon size={'large'} />
         </Icon>
-        <H2 id={i('indexContactId')}>{contact.frontmatter.title}</H2>
-        <Centered dangerouslySetInnerHTML={{ __html: contact.html }} />
+        <H2 id={page.contactId}>{page.contactTitle}</H2>
+        <Centered dangerouslySetInnerHTML={{ __html: page.contactBody.childMarkdownRemark.html }} />
         <Cards>
-          <AuthorContactCard
-            name={anttiKivi.frontmatter.title}
-            job={anttiKivi.frontmatter.job}
-            email={anttiKivi.frontmatter.email}
-            image={getImage(imageAnttiKivi)}
-          />
-          <AuthorContactCard
-            name={gurmannSaini.frontmatter.title}
-            job={gurmannSaini.frontmatter.job}
-            email={gurmannSaini.frontmatter.email}
-            image={getImage(imageGurmannSaini)}
-          />
-          <AuthorContactCard
-            name={anssiMoilanen.frontmatter.title}
-            job={anssiMoilanen.frontmatter.job}
-            email={anssiMoilanen.frontmatter.email}
-            image={getImage(imageAnssiMoilanen)}
-          />
+          {page.contacts.map((contact) => {
+            return <AuthorContactCard key={contact.id} author={contact} />;
+          })}
         </Cards>
       </Section>
     </LayoutIndex>
@@ -170,95 +146,48 @@ const Index = (props) => (
 export default Index;
 
 export const pageQuery = graphql`
-  query IndexQuery($lang: String) {
-    coverMarkdownRemark: markdownRemark(
-      fields: { slug: { eq: "index/cover" }, locale: { eq: $lang } }
-    ) {
-      html
-      frontmatter {
-        title
+  query IndexQuery($locale: String) {
+    contentfulIndexPage(node_locale: { eq: $locale }) {
+      contactId
+      contactTitle
+      introTitle
+      portfolioId
+      portfolioTitle
+      productsId
+      productsTitle
+      storyTitle
+      title
+      contactBody {
+        childMarkdownRemark {
+          html
+        }
       }
-    }
-    storyMarkdownRemark: markdownRemark(
-      fields: { slug: { eq: "index/story" }, locale: { eq: $lang } }
-    ) {
-      html
-      frontmatter {
-        title
-      }
-    }
-    serviceWebsitesMarkdownRemark: markdownRemark(
-      fields: { slug: { eq: "index/websites" }, locale: { eq: $lang } }
-    ) {
-      html
-      frontmatter {
-        title
-      }
-    }
-    serviceDesignMarkdownRemark: markdownRemark(
-      fields: { slug: { eq: "index/design" }, locale: { eq: $lang } }
-    ) {
-      html
-      frontmatter {
-        title
-      }
-    }
-    serviceEventsMarkdownRemark: markdownRemark(
-      fields: { slug: { eq: "index/events" }, locale: { eq: $lang } }
-    ) {
-      html
-      frontmatter {
-        title
-      }
-    }
-    sectionContactMarkdownRemark: markdownRemark(
-      fields: { slug: { eq: "index/contact" }, locale: { eq: $lang } }
-    ) {
-      html
-      frontmatter {
-        title
-      }
-    }
-    authorAnttiKiviMarkdownRemark: markdownRemark(
-      fields: { keySlug: { eq: "/author/antti-kivi" }, locale: { eq: $lang } }
-    ) {
-      frontmatter {
-        title
-        job
+      contacts {
         email
-      }
-    }
-    authorGurmannSainiMarkdownRemark: markdownRemark(
-      fields: { keySlug: { eq: "/author/gurmann-saini" }, locale: { eq: $lang } }
-    ) {
-      frontmatter {
-        title
+        id
         job
-        email
+        name
+        profileImage {
+          gatsbyImageData(width: 200)
+        }
       }
-    }
-    authorAnssiMoilanenMarkdownRemark: markdownRemark(
-      fields: { keySlug: { eq: "/author/anssi-moilanen" }, locale: { eq: $lang } }
-    ) {
-      frontmatter {
+      introBody {
+        childMarkdownRemark {
+          html
+        }
+      }
+      products {
         title
-        job
-        email
+        description {
+          childMarkdownRemark {
+            html
+          }
+        }
       }
-    }
-    imageAnttiKivi: file(relativePath: { eq: "author/antti-kivi.jpg" }) {
-      childImageSharp {
-        gatsbyImageData(width: 200)
-      }
-    }
-    imageGurmannSaini: file(relativePath: { eq: "author/gurmann-saini.jpeg" }) {
-      childImageSharp {
-        gatsbyImageData(width: 200)
-      }
-    }
-    imageAnssiMoilanen: file(relativePath: { eq: "author/anssi-moilanen.jpg" }) {
-      childImageSharp {
-        gatsbyImageData(width: 200)
+      storyBody {
+        childMarkdownRemark {
+          html
+        }
       }
     }
   }