+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.10.0] - 2021-03-16
+
+### Added
+
+- General pages from Contentful.
+- Generation of the slugs of the pages from Contentful by using parent pages.
+
+### Changed
+
+- Front page data to be read from Contentful.
+- Author data to be read from Contentful.
+
+### Removed
+
+- Generation of the search index as long as the transition to Contentful is in progress.
+
 ## [0.9.0] - 2021-03-15
 
 ### Added
@@ -223,7 +239,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.9.0...HEAD
+[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/visiosto/visiosto.fi/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/visiosto/visiosto.fi/compare/v0.8.2...v0.9.0
 [0.8.2]: https://github.com/visiosto/visiosto.fi/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/visiosto/visiosto.fi/compare/v0.8.0...v0.8.1